\name{Banks}
\docType{data}
\alias{Banks}

\title{Volatility-adjusted log returns of the two Banks Citigroup and Bank of America.
}
\description{
A dataset containing the volatility-adjusted log returns of two Banks in the years 2004-2012.
}
\usage{
Banks
}
\format{A list with 9 entries containing the Banks log returns divided by years.
}

\source{Yahoo-Finance.}

