% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CopulaTestTable.R
\name{CopulaTestTable}
\alias{CopulaTestTable}
\title{Applicable dimensions and copula for each test}
\usage{
CopulaTestTable()
}
\value{
A character matrix which consists of dimensions for the 
combination of tests and copula.
}
\description{
\code{\link{CopulaTestTable}} returns a table which shows the applicable
dimensions and copula for each test.
}
\details{
Before performing a gof test with the package on a dataset, it pays out to 
know the implemented copula and dimensions for each test. This
function is dedicated to help finding the applicable tests depending on the 
copula and dimensions available.
}
\examples{

CopulaTestTable()

}
