% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{CryptoCurrencies}
\alias{CryptoCurrencies}
\title{Volatility-adjusted log returns of the two Cryptocurrencies Bitcoin and
Litecoin.}
\format{
A list with 4 entries containing the Cryptocurrencies log returns
divided by years.
}
\source{
CoinMetrics. \url{https://coinmetrics.io/}
}
\description{
A dataset containing the volatility-adjusted log returns of two
Cryptocurrencies in the years 2015-2018.
}
