% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Test4Copula.R
\name{gofWhich}
\alias{gofWhich}
\title{The function \code{gofWhich} was renamed to \code{gofTest4Copula}. Please
use \code{gofTest4Copula}.}
\usage{
gofWhich(copula = NULL, d = 2)
}
\arguments{
\item{copula}{The copula to test for. Possible are the copulae
\code{"normal"}, \code{"t"}, \code{"clayton"}, \code{"gumbel"}, 
\code{"frank"}, \code{"joe"}, \code{"amh"}, \code{"galambos"}, 
\code{"huslerReiss"}, \code{"tawn"}, \code{"tev"}, \code{"fgm"} and 
\code{"plackett"}. For the default \code{NULL}, all tests in the package are
returned.}

\item{d}{The dimension to search for.}
}
\value{
A character vector which consists of the names of the tests.
}
\description{
Please see and use the function \code{\link{gofTest4Copula}}.
}
