% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Copula4Test.R
\name{gofWhichCopula}
\alias{gofWhichCopula}
\title{The function \code{gofWhichCopula} was renamed to \code{gofCopula4Test}.
Please use \code{gofCopula4Test}.}
\usage{
gofWhichCopula(test)
}
\arguments{
\item{test}{The test to search for copula.}
}
\value{
A character vector which consists of the names of the copula.
}
\description{
Please see and use the function \code{\link{gofCopula4Test}}.
}
