\name{IndexReturns}
\docType{data}
\alias{IndexReturns}

\title{Log returns of european stock indices.
}
\description{
A dataset containing the log returns of 4 european stock indices in the time period 1991 - 1998.
}
\usage{
IndexReturns
}
\format{A matrix with 1859 rows and 4 columns
}

\source{The raw data are part of the R datasets and were kindly provided by 
Erste Bank AG, Vienna, Austria.}

