\name{IndexReturns3D}
\docType{data}
\alias{IndexReturns3D}

\title{Log returns of european stock indices
}
\description{
A dataset containing the log returns of three european stock indices in the year 1998.
}
\usage{
IndexReturns3D
}
\format{A matrix with 200 rows and 3 columns
}

\source{The raw data are part of the R datasets and were kindly provided by 
Erste Bank AG, Vienna, Austria.}

