% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_gofCOP_print.R
\name{print.gofCOP}
\alias{print.gofCOP}
\title{Printing function for objects of class gofCOP}
\usage{
\method{print}{gofCOP}(x, ...)
}
\arguments{
\item{x}{An object of \code{class} gofCOP.}

\item{...}{Further arguments to be passed over. Currently non.}
}
\value{
None
}
\description{
\code{\link{print.gofCOP}} prints the values of an object of class gofCOP.
}
