% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_est_time.R
\name{print.goftime}
\alias{print.goftime}
\title{Printing function for an object of class goftime. goftime objects are created
by the function gofCheckTime.}
\usage{
\method{print}{goftime}(x, ...)
}
\arguments{
\item{x}{An object of class goftime}

\item{...}{Additional arguments passed to print}
}
\description{
Printing function for an object of class goftime. goftime objects are created
by the function gofCheckTime.
}
