\name{print.gofCOP}
\alias{print.gofCOP}

\title{Printing function for objects of class gofCOP
}
\description{
\code{\link{print.gofCOP}} prints the values of an object of class gofCOP.
}
\usage{
\method{print}{gofCOP}(x, ...)
}
\arguments{
\item{x}{
An object of \code{class} gofCOP.}
\item{...}{
Further arguments to be passed over. Currently non.
}

}

\value{None}


