\name{gofRn}
\alias{gofRn}

\title{Removed due to inconsistencies with the remaining tests
}

\description{
This test was removed due to inconsistencies with the remaining tests. If you still like to run it, please see the package copula.
}
\usage{
gofRn(...)
}
\arguments{
\item{...}{
Parameters previously used in gofRn.
}
}