% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Banks}
\alias{Banks}
\title{Volatility-adjusted log returns of the two Banks Citigroup and Bank of
America.}
\format{
A list with 9 entries containing the Banks log returns divided by
years.
}
\source{
Yahoo-Finance.
}
\description{
A dataset containing the volatility-adjusted log returns of two Banks in the
years 2004-2012.
}
