\name{CryptoCurrencies}
\docType{data}
\alias{CryptoCurrencies}

\title{Volatility-adjusted log returns of the two Cryptocurrencies Bitcoin and Litecoin.
}
\description{
A dataset containing the Volatility-adjusted log returns of two Cryptocurrencies in the years 2015-2018.
}
\usage{
CryptoCurrencies
}
\format{A list with 4 entries containing the Cryptocurrencies log returns divided by years.
}

\source{CoinMetrics.}

