\name{IndexReturns2D}
\docType{data}
\alias{IndexReturns2D}

\title{Log returns of european stock indices
}
\description{
A dataset containing the log returns of two european stock indices in the year 1998.
}
\usage{
IndexReturns2D
}
\format{A matrix with 100 rows and 2 columns
}

\source{The raw data are part of the R datasets and were kindly provided by 
Erste Bank AG, Vienna, Austria.}

