% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{IndexReturns3D}
\alias{IndexReturns3D}
\title{Log returns of european stock indices}
\format{
A matrix with 200 rows and 3 columns
}
\source{
The raw data are part of the R datasets and were kindly provided by
Erste Bank AG, Vienna, Austria.
}
\description{
A dataset containing the log returns of three european stock indices in the
year 1998.
}
